refactor: 검색 기능 훅 추가 및 App.tsx에서 검색 로직 통합
- useSearch 훅을 추가하여 검색 관련 상태 및 로직을 분리
- App.tsx에서 검색 상태 관리 및 debouncedQuery 로직을 useSearch로 통합
- 불필요한 코드 삭제 및 코드 정리
