+++ src/basic/hooks/search/useSearch.ts
@@ -1,0 +1,19 @@
+import { useEffect, useState } from 'react';
+
+export const useSearch = () => {
+  const [query, setQuery] = useState('');
+  const [debouncedQuery, setDebouncedQuery] = useState('');
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setDebouncedQuery(query);
+    }, 500);
+    return () => clearTimeout(timer);
+  }, [query]);
+
+  return {
+    query,
+    setQuery,
+    debouncedQuery,
+  };
+};
